FastAPI/requirements.txt

fastapi==0.95.1
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
requests==2.30.0
tensorflow==2.12.0 
uvicorn==0.22.0
